requirements.txt
fastapi
uvicorn
pyairtable
python-dotenv
pydantic
httpx
httpie
pyjwt
python-jose
passlib
python-dotenv
numpy
pandas
PyPDF2
python-dotenv
transformers
sentence-transformers
huggingface-hub  
argparse
langchain
openai
gradio
chainlit
faiss-cpu
chainlit
faiss-cpu
nltk
spacy
argon2-cffi